requirements.txt

fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
requests==2.32.3
aiohttp==3.10.0
numpy==1.26.0
redis==5.0.8
aioredis==2.0.1
pyjwt==2.9.0
cryptography==43.0.0
